POC permissions endpoint (ref #600)
